+++ lib/Macaroons/Macaroon.php
@@ -31,6 +31,18 @@ class Macaroon
     return strtolower( Utils::hexlify( $this->signature ) );
   }
 
+  public function setSignature($signature)
+  {
+    if (!$signature)
+      throw new \InvalidArgumentException('Must supply updated signature');
+    $this->signature = $signature;
+  }
+
+  public function setCaveats(Array $caveats)
+  {
+    $this->caveats = $caveats;
+  }
+
   public function addFirstPartyCaveat($predicate)
   {
     $this->caveats[] = new Caveat($predicate);
@@ -52,5 +64,83 @@ class Macaroon
   private function initialSignature($key, $identifier)
   {
     return Utils::hmac( Utils::generateDerivedKey($key), $identifier);
+  }
+
+  // TODO: Move these into a separate object
+  public function serialize()
+  {
+    $p = new Packet();
+    $s = $p->packetize(
+                        array(
+                          'location' => $this->location,
+                          'identifier' => $this->identifier
+                        )
+                      );
+    foreach ($this->caveats as $caveat)
+    {
+      $p = new Packet();
+      $s = $s . $p->packetize(
+                                array(
+                                  'vid' => $caveat->getVerificationId(),
+                                  'cl' => $caveat->getCaveatLocation()
+                                )
+                              );
+    }
+    $p = new Packet();
+    $s = $s . $p->packetize(array('signature' => $this->signature));
+    return Utils::base64_url_encode($s);
+  }
+
+  public static function deserialize($serialized)
+  {
+    $location = NULL;
+    $identifier = NULL;
+    $signature = NULL;
+    $caveats = array();
+    $decoded = Utils::base64_url_decode($serialized);
+
+    $index = 0;
+
+    while ($index < strlen($decoded))
+    {
+      // TOOD: Replace 4 with PACKET_PREFIX_LENGTH
+      $packetLength = hexdec(substr($decoded, $index, 4));
+      $packetDataStart = $index + 4;
+      $strippedPacket = substr($decoded, $packetDataStart, strpos($decoded, "\n", $index) - $packetDataStart);
+      $packet = new Packet();
+      $packet = $packet->decode($strippedPacket);
+
+      switch($packet->getKey())
+      {
+        case 'location':
+          $location = $packet->getData();
+        break;
+        case 'identifier':
+          $identifier = $packet->getData();
+        break;
+        case 'signature':
+          $signature = $packet->getData();
+        break;
+        case 'cid':
+          $caveat = new Caveat($packet->getData());
+        break;
+        case 'vid':
+          $caveat = $caveats[ count($caveats) - 1 ];
+          $caveat->setVerificationId($packet->getData());
+        break;
+        case 'cl':
+          $caveat = $caveats[ count($caveats) - 1 ];
+          $caveat->setCaveatLocation($packet->getData());
+        break;
+        default:
+          throw new \DomainException('Invalid key in binary macaroon. Macaroon may be corrupted.');
+        break;
+      }
+      $index = $index + $packetLength;
+    }
+    $m = new Macaroon('no_key', $identifier, $location);
+    $m->setCaveats($caveats);
+    $m->setSignature($signature);
+    return $m;
   }
 }