+++ tests/MacaroonTest.php
@@ -21,6 +21,19 @@ class MacaroonTest extends \PHPUnit_Framework_TestCase
     $this->assertEquals('e3d9e02908526c4c0039ae15114115d97fdd68bf2ba379b342aaf0f617d0552f', $this->m->getSignature());
   }
 
+  public function testSerialize()
+  {
+    $this->assertEquals('MDAxZGxvY2F0aW9uIGh0dHBzOi8vbXliYW5rLwowMDI2aWRlbnRpZmllciB3ZSB1c2VkIG91ciBzZWNyZXQga2V5CjAwMmZzaWduYXR1cmUg49ngKQhSbEwAOa4VEUEV2X_daL8ro3mzQqrw9hfQVS8K', $this->m->serialize());
+  }
+
+  public function testDeserialize()
+  {
+    $m = Macaroon::deserialize($this->m->serialize());
+    $this->assertEquals($this->m->getSignature(), $m->getSignature());
+    $this->assertEquals($this->m->getIdentifier(), $m->getIdentifier());
+    $this->assertEquals($this->m->getLocation(), $m->getLocation());
+  }
+
   public function testFirstPartyCaveat()
   {
     $this->m->addFirstPartyCaveat('test = caveat');